Implement the death animation
Update all current functions and create some new ones in order to implement
the death animation with a big reorganise. Create another folder that contains
the functions to deal with this animation.
